requirements.txt
apache-airflow==2.8.1
apache-airflow-providers-mysql
apache-airflow-providers-common-sql
